+++ objectbox/lib/src/native/store.dart
@@ -229,13 +229,19 @@ class Store implements Finalizable {
       : _closesNativeStore = true,
         _absoluteDirectoryPath = _safeAbsoluteDirectoryPath(directory) {
     try {
-      if (Platform.isMacOS && macosApplicationGroup != null) {
+      if (macosApplicationGroup != null) {
+        final isGroupEmpty = macosApplicationGroup.isEmpty;
+        // Docs don't explicitly require a trailing slash, so add one if missing
         if (!macosApplicationGroup.endsWith('/')) {
           macosApplicationGroup += '/';
         }
-        if (macosApplicationGroup.length > 20) {
-          ArgumentError.value(macosApplicationGroup, 'macosApplicationGroup',
-              'Must be at most 20 characters long');
+        // The database library would check for length, but refers to 'prefix'
+        // instead of the parameter name used here. So duplicate the checks.
+        if (isGroupEmpty || macosApplicationGroup.length > 20) {
+          throw ArgumentError.value(
+              macosApplicationGroup,
+              'macosApplicationGroup',
+              'Must be at least 1 and at most 19 characters long');
         }
         // This is required to enable additional interprocess communication
         // (IPC) in sandboxed apps (https://developer.apple.com/documentation/xcode/configuring-app-groups),
@@ -243,7 +249,10 @@ class Store implements Finalizable {
         // library. macOS requires that semaphore names are prefixed with an
         // application group ID.
         // See the constructor docs for more details.
-        withNativeString(macosApplicationGroup, C.posix_sem_prefix_set);
+        // Note: calling this on all platforms is fine, it will be a no-op if
+        // not supported.
+        checkObx(
+            withNativeString(macosApplicationGroup, C.posix_sem_prefix_set));
       }
       _checkStoreDirectoryNotOpen();
       final model = Model(modelDefinition.model);